Shelby/breakup autobuild (#1138)
* Break up the autobuild

To decrease the time the autobuilds take in AppVeyor, we want to break
up the tests so that they run in parallel. With premium version in
AppVeyor, we can run two jobs in parallel. The order of these tests
does matter. We do not want two long running test categories running
on the same virtual machine.

* Break up the autobuild

Changed order of scheduled jobs to see how long unit
tests run as a secondary job.

* Break up the autobuild

Changed order of scheduled jobs to see how long unit
tests run as a secondary job.

* Wrapped some init commands with timing statements

In order to help identify with decreasing init time for appveyor
builds, added some timing statements around certain init commands.
Also merged some of the testing jobs together.

* Modified init timing command

Previous approach using Out-Host failed due to Out-Host not being
recognized as an internal or external command. Attempting a
different approach that uses Measure-Command and gets the total
seconds elapsed.

* Added timing statements as part of a powershell

* Added fast_finish for failing second job if first job fails

In the event that the second job hasn't already launched in a VM,
fail it if the first job failed. Modified the timing statement
around tests.

* Changed time duration logic for launching tests

Moved npm run %vscode$ into its own powershell script to avoid
errors.

* Modification for appveyor timing

* Modified timing output line

* Added rollup time statements for appveyor

* Added in changes from develop for autobuild failures

* Changed stopwatch reset and start logic

* Change xunit and junit-custom.xml name generation

When unit tests and integration tests were running for the same package we
were seeing that the .xml test files that were generated were overwriting
one another. This is because the results of both tests were being piped
to the same files. Each test type should now be piped to its own xunit
and junit-custom.xml file.

* Break up the autobuild

To decrease the time the autobuilds take in AppVeyor, we want to break
up the tests so that they run in parallel. With premium version in
AppVeyor, we can run two jobs in parallel. The order of these tests
does matter. We do not want two long running test categories running
on the same virtual machine.

* Change xunit and junit-custom.xml name generation

When unit tests and integration tests were running for the same package we
were seeing that the .xml test files that were generated were overwriting
one another. This is because the results of both tests were being piped
to the same files. Each test type should now be piped to its own xunit
and junit-custom.xml file.

* Remove config for failing job if other job fails

We want to be able to see the results of other jobs even if
the current job fails.

* Removed unneeded scripts that were added for testing