Update bug template to ask for SF EXTENSION version (#3361)
we often have to ask a follow-up question to see which extension version they are on.  Changing the template to ask that upfront.  if we need to also know the VS Code version, we will ask in follow-ups and consider adding that to the template if we have to ask a lot

Co-authored-by: Jag Jayaprakash <58611665+jag-sfdc@users.noreply.github.com>
Co-authored-by: AnanyaJha <a.jha@salesforce.com>